Prevent the gallery view keystroke from stacking up.
The CTRL-ALT-G shortcut sometimes takes awhile, because different
monitors are at different DPI/scales and the window takes some time to
settle to its new position.

This results in confusion, because it's easy to press the shortcut again
before it has completed.

This commit ignores all shortcut keypresses until half a second has
elapsed from the completion of the previous keypress.