Add new shortcuts to move gallery to specified monitors
These are coincident with the keys in OBS configuration to select
different modes, in an attempt to increase ease of use:

CTRL-ALT-C:  In OBS, selects classroom view.  This also moves the
gallery to the tertiary monitor.

CTRL-ALT-L: In OBS, selects laptop camera.  This also moves the gallery
to the laptop screen.

It is still possible to use CTRL-ALT-G to get the gallery wherever you'd
prefer.